Move logic into context
