+++ src/contexts/BattleClientContext.tsx
@@ -3,12 +3,24 @@
  * @Brief Global battle client context.
  */
 
-import { BattleClient, IBattleClient } from "combatcritters-ts";
-import React, { createContext, useContext, useState, ReactNode } from "react";
+import {
+  BattleClient,
+  IBattleClient,
+  IMatchStateObserver,
+} from "combatcritters-ts";
+import React, {
+  createContext,
+  useContext,
+  useState,
+  ReactNode,
+  useEffect,
+} from "react";
+import { useNavigate } from "react-router-dom";
+import { useBattleState } from "./BattleStateContext";
 
 interface BattleClientType {
   battleClient: IBattleClient | null;
-  refreshClient: () => void;
+  fetchBattleClient: () => void;
 }
 
 const BattleClientContext = createContext<BattleClientType | undefined>(
@@ -16,31 +28,63 @@ const BattleClientContext = createContext<BattleClientType | undefined>(
 );
 
 export const BattleClientProvider = ({ children }: { children: ReactNode }) => {
-  const battleRoot: string =
+  const BATTLE_ROOT: string =
     process.env.REACT_APP_SOCKET ?? "ws://api.combatcritters.ca:4000/ws";
 
   const [battleClient, setBattleClient] = useState<IBattleClient | null>(null);
+  const { battleStateObserver } = useBattleState();
+
+  const navigate = useNavigate();
 
   /**
    * Get a new battle client
    */
-  const refreshClient = async () => {
+  const fetchBattleClient = async () => {
     const refresh = async () => {
       try {
-        const bClient = await BattleClient.getClient(battleRoot);
+        const bClient = await BattleClient.getClient(BATTLE_ROOT);
         setBattleClient(bClient);
       } catch (error) {
-        console.error("Failed to refresh battle client:" + error);
+        console.error("Failed to fetch battle client:" + error);
       }
     };
     refresh();
   };
 
+  /**
+   * Initialize the client once ready
+   */
+  useEffect(() => {
+    const initialize = async () => {
+      try {
+        battleClient?.setBattleStateObserver(battleStateObserver);
+        battleClient?.onStopped(() => {
+          navigate("/home");
+        });
+        battleClient?.setMatchStateObserver(new MatchObserver());
+      } catch (error) {
+        console.error("Failed to initialize battle client:" + error);
+      }
+    };
+    initialize();
+  }, [battleClient]);
+
+  /**
+   * The match observer attached to the battle client
+   */
+  class MatchObserver implements IMatchStateObserver {
+    gameFound(opponent: string): void {
+      console.log("Game found against " + opponent);
+      navigate("/battle");
+    }
+    matchEnded(_: string): void {
+      console.log("Match Ended");
+      navigate("/home");
+    }
+  }
 
   return (
-    <BattleClientContext.Provider
-      value={{ battleClient, refreshClient  }}
-    >
+    <BattleClientContext.Provider value={{ battleClient, fetchBattleClient }}>
       {children}
     </BattleClientContext.Provider>
   );
@@ -49,7 +93,9 @@ export const BattleClientProvider = ({ children }: { children: ReactNode }) => {
 export const useBattleClient = (): BattleClientType => {
   const context = useContext(BattleClientContext);
   if (context === undefined) {
-    throw new Error("useBattleClient must be used within a BattleClientProvider");
+    throw new Error(
+      "useBattleClient must be used within a BattleClientProvider"
+    );
   }
   return context;
 };
